requirements (2).txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2 